Moar Java updates

+++ java/usbsid-usb-driver-library-java/src/main/java/usbsid/IUSBSID.java
@@ -8,9 +8,9 @@ public interface IUSBSID {
 
   void USBSID_reset(byte volume);
 
-  void USBSID_clkdwrite(byte addr, byte data, long cycles);
+  void USBSID_clkdwrite(byte addr, byte data, short cycles);
 
-  void USBSID_writeclkdbuffer(byte addr, byte data, long cycles);
+  void USBSID_writeclkdbuffer(byte addr, byte data, short cycles);
 
   void USBSID_setflush();
